+++ grammar/en/src/training.rs
@@ -47,7 +47,7 @@ pub fn examples_finance(v: &mut Vec<::rustling::train::Example<Dimension>>) {
 
 // TODO: Sort out and split by datetime subtype
 pub fn examples_datetime(v: &mut Vec<::rustling::train::Example<Dimension>>) {
-    let c = ResolverContext::new(Interval::starting_at(Moment(Local.ymd(2013, 2, 12).and_hms(4, 30, 0)), Grain::Second));
+    let c = ResolverContext::for_reference(Interval::starting_at(Moment(Local.ymd(2013, 2, 12).and_hms(4, 30, 0)), Grain::Second));
     example!(v, check_moment!(c, [2013, 2, 12, 4, 30, 0]), "now", "right now", "just now", "at this time");
     example!(v, check_moment!(c, [2013, 2, 12]), "today");
     example!(v, check_moment!(c, [2013, 2, 11]), "yesterday");