+++ grammar/it/src/training.rs
@@ -4,7 +4,7 @@ use rustling_ontology_values::dimension::*;
 use rustling_ontology_values::ResolverContext;
 
 pub fn examples_datetime(v: &mut Vec<::rustling::train::Example<Dimension>>) {
-    let c = ResolverContext::new(Interval::starting_at(Moment(Local.ymd(2013, 2, 12).and_hms(4, 30, 0)), Grain::Second));
+    let c = ResolverContext::for_reference(Interval::starting_at(Moment(Local.ymd(2013, 2, 12).and_hms(4, 30, 0)), Grain::Second));
     // Days
     example!(v, check_moment!(c, [2013, 2, 12, 4, 30, 00]), "ora", "adesso", "in questo momento esatto", "in questo preciso istante");
     example!(v, check_moment!(c, [2013, 2, 12]), "oggi", "in questo momento", "in questa giornata");